fix: use PRIu32 in all stepper logs

+++ main/main.c
@@ -148,7 +148,7 @@ static void motor_setup(void)
     ESP_LOGI(TAG, "STEP=%d DIR=%d EN=%d", STEP_GPIO, DIR_GPIO, EN_GPIO);
     ESP_LOGI(TAG, "Enable pin is %s", EN_ACTIVE_LOW ? "active-low" : "active-high");
     ESP_LOGI(TAG, "PWM step frequency=%" PRIu32 " Hz (%" PRIu32 " us period), resolution=%" PRIu32 " bits", freq_hz, (uint32_t)STEP_PERIOD_US, resolution_bits);
-    ESP_LOGI(TAG, "Move=%u ms pause=%u ms", MOVE_TIME_MS, PAUSE_MS);
+    ESP_LOGI(TAG, "Move=%" PRIu32 " ms pause=%" PRIu32 " ms", (uint32_t)MOVE_TIME_MS, (uint32_t)PAUSE_MS);
 }
 
 void app_main(void)
